CT-31 First blueprint fsd is done
