--- client/web/src/components/Footer/footer.module.css
@@ -1,11 +1,0 @@
-.footer {
-    height: 80px;
-    display: flex;
-    justify-content: space-around;
-    font-family: roboto;
-    align-items: center;
-}
-
-.footer__logo a {
-    text-decoration: none;
-}
