--- client/web/src/components/Map/Map.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import cl from './map.module.scss';
-
-const Map = () => {
-    return (
-        <div>
-            {/* <section id='map-test' className={cl['map']}></section> */}
-        </div>
-    );
-};
-
-export default Map;
