--- client/web/src/components/Map/map.module.scss
@@ -1,5 +1,0 @@
-.map {
-    background-color: darkgrey;
-    width: 100%;
-    height: auto;
-}
