--- client/web/src/components/Sidebar/sidebar.module.scss
@@ -1,7 +1,0 @@
-.sidebar {
-    width: 200px;
-    height: auto;
-    background-color: lightgray;
-    padding: 10px;
-    text-align: start;
-}
