+++ client/web/src/shared/api/store/hooks/useActions.ts
@@ -1,6 +1,6 @@
 import { useDispatch } from "react-redux";
 import { bindActionCreators } from "redux";
-import ActionCreators from '../shared/api/store/action-creators';
+import ActionCreators from "../action-creators";
 
 export const useActions = () => {
     const dispatch = useDispatch();
