+++ client/web/src/widgets/AuthBar/AuthBar.tsx
