+++ client/web/src/widgets/Footer/styles.module.css
@@ -1,0 +1,11 @@
+.footer {
+    height: 80px;
+    display: flex;
+    justify-content: space-around;
+    font-family: roboto;
+    align-items: center;
+}
+
+.footer__logo a {
+    text-decoration: none;
+}
