+++ client/web/src/widgets/Map/Map.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from './styles.module.scss';
+
+const Map = () => {
+    return (
+        <div>
+            {/* <section id='map-test' className={styles['map']}></section> */}
+        </div>
+    );
+};
+
+export default Map;
