+++ client/web/src/widgets/Map/styles.module.scss
@@ -1,0 +1,5 @@
+.map {
+    background-color: darkgrey;
+    width: 100%;
+    height: auto;
+}
