+++ client/web/src/widgets/Sidebar/styles.module.scss
@@ -1,0 +1,7 @@
+.sidebar {
+    width: 200px;
+    height: auto;
+    background-color: lightgray;
+    padding: 10px;
+    text-align: start;
+}
